tests: Refactors API and adds integration tests
Improves API models by adding validation and using enums for the `tipo` field.  This ensures data integrity and clarity.
Updates the response model to use `pydantic.ConfigDict` for better compatibility and removes redundant JSON processing.

Adds integration tests to verify API functionality and data integrity, including tests for validation rules.  The tests use a dedicated test database to avoid affecting the development data.  This increases the reliability of the code and helps prevent regressions.

+++ contas_a_pagar_e_receber/models/conta_a_pagar_receber_model.py
@@ -11,5 +11,5 @@ class ContaPagarReceber(Base):
 
     id: int = Column(Integer, primary_key=True, autoincrement=True)
     descricao: str = Column(String(30))
-    valor: Decimal = Column(Numeric)
+    valor: float = Column(Numeric(15, 2))
     tipo: str = Column(String(30))
